requirements.txt
streamlit
pydub
openai
pillow
numpy
pandas
python-dotenv